ci: refactoring
