+++ declarations.d.ts
@@ -1,0 +1,5 @@
+declare module "*.md" {
+  const attributes: any;
+  const html: string;
+  export { attributes, html };
+}
