+++ index.html
@@ -15,9 +15,7 @@
   <body>
     <div id="pixiContainer"><canvas id="pixiCanvas"></canvas></div>
     <div class="options">
-      <h3>skins</h3>
       <section id="optionsSkins"></section>
-      <h3>animations</h3>
       <section id="optionsAnimations"></section>
     </div>
     <div id="controlsContainer"></div>
