+++ styles.css
@@ -299,30 +299,36 @@ div#stats {
   user-select: none;
 }
 
-.options {
+.options section {
+  margin-bottom: 20px;
+}
+
+.options-wrapper {
   display: flex;
   flex-direction: column;
   align-items: center;
-  text-align: center;
 }
 
 .options h3 {
   text-align: center;
+  margin-bottom: 10px;
 }
 
-.options #optionsSkins {
+.buttons-container {
   display: grid;
   grid-template-columns: repeat(auto-fit, minmax(150px, 1fr));
+  gap: 4px;
+  justify-content: center;
+}
+
+.options #optionsSkins .buttons-container {
   max-width: calc((150px * 2) + 8px);
 }
 
-.options #optionsAnimations {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(150px, 1fr));
+.options #optionsAnimations .buttons-container {
   max-width: calc((150px * 3) + 8px);
 }
 
 .options section button {
   width: 150px;
-  margin: 2px;
 }
