+++ vite.config.ts
@@ -1,6 +1,6 @@
 import { defineConfig, loadEnv } from "vite";
 import { resolve } from "path";
-import mdPlugin, { Mode } from 'vite-plugin-markdown';
+import mdPlugin, { Mode } from "vite-plugin-markdown";
 
 export default defineConfig(async ({ mode }) => {
   return {
@@ -9,7 +9,7 @@ export default defineConfig(async ({ mode }) => {
       port: 8080,
     },
 
-    plugins: [mdPlugin({mode: Mode.HTML as any})],
+    plugins: [mdPlugin({ mode: Mode.HTML as any })],
     build: {
       minify: false,
       minifyIdentifiers: false,
